requirements.txt
aniso8601==1.0.0
Flask==0.10.1
flask-appconfig==0.9.1
Flask-Auth==0.85
Flask-Markdown==0.3
Flask-RESTful==0.3.3
Flask-SQLAlchemy==2.1
gunicorn==19.3.0
itsdangerous==0.24
Jinja2==2.7.3
Markdown==2.6.5
MarkupSafe==0.23
mongo==0.2.0
MongoAlchemy==0.19
nltk==3.0.2
pymongo==2.8.1
python-cjson==1.1.0
pytz==2015.4
six==1.9.0
SQLAlchemy==1.0.11
Werkzeug==0.10.4
wheel==0.24.0
